refactor and dedup code between apple and spotify metadata caches
