+++ listenbrainz/labs_api/labs/api/apple/__init__.py
@@ -1,12 +1,7 @@
-import uuid
 from typing import Optional
 
-from pydantic import BaseModel
+from listenbrainz.labs_api.labs.api.metadata_index import BaseMetadataIndexOutput
 
 
-class AppleMusicIdFromMBIDOutput(BaseModel):
-    recording_mbid: Optional[uuid.UUID]
-    artist_name: Optional[str]
-    release_name: Optional[str]
-    track_name: Optional[str]
-    apple_track_ids: Optional[list[str]]
+class AppleMusicIdFromMBIDOutput(BaseMetadataIndexOutput):
+    apple_music_track_ids: Optional[list[str]]
