+++ listenbrainz/labs_api/labs/api/spotify/__init__.py
@@ -1,12 +1,7 @@
-import uuid
 from typing import Optional
 
-from pydantic import BaseModel
+from listenbrainz.labs_api.labs.api.metadata_index import BaseMetadataIndexOutput
 
 
-class SpotifyIdFromMBIDOutput(BaseModel):
-    recording_mbid: Optional[uuid.UUID]
-    artist_name: Optional[str]
-    release_name: Optional[str]
-    track_name: Optional[str]
+class SpotifyIdFromMBIDOutput(BaseMetadataIndexOutput):
     spotify_track_ids: Optional[list[str]]
