+++ listenbrainz/labs_api/labs/api/spotify/spotify_metadata_lookup.py
@@ -1,24 +1,16 @@
-from datasethoster import Query
-from pydantic import BaseModel
-
+from listenbrainz.labs_api.labs.api.metadata_index.metadata_index_from_metadata_lookup import \
+    MetadataIndexFromMetadataQuery
 from listenbrainz.labs_api.labs.api.spotify import SpotifyIdFromMBIDOutput
-from listenbrainz.labs_api.labs.api.utils import lookup_using_metadata
 
 
-class SpotifyIdFromMetadataInput(BaseModel):
-    artist_name: str
-    release_name: str
-    track_name: str
-
-
-class SpotifyIdFromMetadataQuery(Query):
+class SpotifyIdFromMetadataQuery(MetadataIndexFromMetadataQuery):
     """ Query to lookup spotify track ids using artist name, release name and track name. """
+
+    def __init__(self):
+        super().__init__("spotify")
 
     def names(self):
         return "spotify-id-from-metadata", "Spotify Track ID Lookup using metadata"
-
-    def inputs(self):
-        return SpotifyIdFromMetadataInput
 
     def introduction(self):
         return """Given the name of an artist, the name of a release and the name of a recording (track)
@@ -26,7 +18,3 @@ class SpotifyIdFromMetadataQuery(Query):
 
     def outputs(self):
         return SpotifyIdFromMBIDOutput
-
-    def fetch(self, params, source, offset=-1, count=-1):
-        data = [p.dict() for p in params]
-        return lookup_using_metadata(data, 'spotify')
